Emil-8/Emil-task8-task-202411071115/gen_src/Simulation.java
import java.util.List;

public class Simulation {
    private List<CelestialBody> bodies;
    private final double G = 6.67430e-11;

    public Simulation(List<CelestialBody> bodies) {
        // Constructor implementation
    }

    public void calculateGravitationalForces(double timeStep) {
        // Method to calculate gravitational forces
    }

    private double[] computeForce(CelestialBody a, CelestialBody b) {
        // Method to compute force between two celestial bodies
        return new double[]{};
    }

    private void updateVelocityAndPosition(CelestialBody body, double[] netForce, double timeStep) {
        // Method to update velocity and position of a celestial body
    }

    public void simulate(double totalTime, double timeStep) {
        // Method to simulate the system over a given total time with specified time steps
    }

    private void printSimulationState(double currentTime) {
        // Method to print the current state of the simulation
    }
}